# packages/contracts/openapi/plotify-chat.v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetaChange": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/MetaValue"
          }
        },
        "required": [
          "value",
          "field"
        ],
        "title": "MetaChange",
        "type": "object"
      },
      "MetaContact": {
        "properties": {
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          },
          "wa_id": {
            "title": "Wa Id",
            "type": "string"
          }
        },
        "required": [
          "profile",
          "wa_id"
        ],
        "title": "MetaContact",
        "type": "object"
      },
      "MetaEntry": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/MetaChange"
            },
            "title": "Changes",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "changes"
        ],
        "title": "MetaEntry",
        "type": "object"
      },
      "MetaMessage": {
        "properties": {
          "from": {
            "title": "From",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetaText"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "type": {
            "default": "unknown",
            "enum": [
              "text",
              "image",
              "audio",
              "document",
              "interactive",
              "button",
              "unknown"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "from",
          "timestamp"
        ],
        "title": "MetaMessage",
        "type": "object"
      },
      "MetaText": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "MetaText",
        "type": "object"
      },
      "MetaValue": {
        "properties": {
          "contacts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MetaContact"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacts"
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MetaMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          },
          "messaging_product": {
            "title": "Messaging Product",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "statuses": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statuses"
          }
        },
        "required": [
          "messaging_product",
          "metadata"
        ],
        "title": "MetaValue",
        "type": "object"
      },
      "MetaWebhookPayload": {
        "properties": {
          "entry": {
            "items": {
              "$ref": "#/components/schemas/MetaEntry"
            },
            "title": "Entry",
            "type": "array"
          },
          "object": {
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "object",
          "entry"
        ],
        "title": "MetaWebhookPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Plotify Messaging Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/webhook/meta": {
      "get": {
        "description": "Endpoint de Verificación de Meta (Hub Challenge).\nObligatorio para configurar el Webhook en la app de Meta.\nVerifica que el hub.verify_token coincida con el \".env\" y retorna el hub.challenge.",
        "operationId": "verify_meta_webhook_api_v1_webhook_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Verify Meta Webhook",
        "tags": [
          "webhooks"
        ]
      },
      "post": {
        "description": "Endpoint para recibir Webhooks de Meta (Mensajes entrantes y estados).\nRegla de Oro: Siempre retornar 200 OK de inmediato a Meta.\nTodo el procesamiento pesado va asíncrono vía la cola Redis (arq).",
        "operationId": "receive_meta_webhook_api_v1_webhook_meta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetaWebhookPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receive Meta Webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/v1/webhook/telegram/{org_id}": {
      "post": {
        "description": "Endpoint para recibir Webhooks de Telegram con ruteo por organización.\nManeja dos tipos de eventos:\n1. Mensajes de texto → se encolan para el agente LangGraph.\n2. callback_query (botones inline) → se encolan para procesar decisiones de aprobación.",
        "operationId": "receive_telegram_webhook_api_v1_webhook_telegram__org_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receive Telegram Webhook",
        "tags": [
          "webhooks"
        ]
      }
    }
  }
}